+++ crates/miden-objects/src/account/component/mod.rs
@@ -207,7 +207,7 @@ impl AccountComponent {
         let mut procedures = Vec::new();
         for module in self.library.module_infos() {
             for (_, procedure_info) in module.procedures() {
-                let is_auth = procedure_info.name.contains("auth__");
+                let is_auth = procedure_info.name.starts_with("auth_");
                 procedures.push((procedure_info.digest, is_auth));
             }
         }