+++ crates/miden-testing/src/kernel_tests/tx/test_tx.rs
@@ -458,7 +458,7 @@ fn user_code_can_abort_transaction_with_summary() -> anyhow::Result<()> {
       use.miden::tx
       #! Inputs:  [AUTH_ARGS, pad(12)]
       #! Outputs: [pad(16)]
-      export.auth__abort_tx
+      export.auth_abort_tx
           dropw
           # => [pad(16)]